refactor: change pathTrie per-file to per-directory (#1931)
Now the `pathTrie` aligns with the directory tree, not full fs tree
which includes files. Aligning with the directory tree makes more sense
to me, and I think will make it simpler to do things like load BUILDs
concurrently while doing the fs-traversal (and persisting that in the
`pathTrie`) etc.

**What type of PR is this?**

> Other

**What package or component does this PR mostly affect?**

> all

**What does this PR do? Why is it needed?**

Performance + simplifying things for future changes.