Support load fixing on struct members (dotexpr) instead of just callexpr (#1972)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

> Bug fix

**What package or component does this PR mostly affect?**

> For example:
>
> all

**What does this PR do? Why is it needed?**

The changes introduced by #1164 largely support what I need – automatic
loading of non-rule symbols from known loads – except it only supports
`CallExpr`. For imported struct fields, these are not automatically
picked up, and the ergonomics of getting them loaded in Gazelle has been
tricky for me. I am hoping this is a reasonable fix to this.

**Which issues(s) does this PR fix?**

Fixes #1971 

**Other notes for review**

I haven't written Go in a while so please forgive the mess.