[Proto] Make protoFileInfo method public (#1998)
Summary:
In this commit, we make the protoFileInfo method public so that users
interested in parsing the proto files can use the method directly.

<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

Other

**What package or component does this PR mostly affect?**


language/proto


**What does this PR do? Why is it needed?**

This PR makes a previously private method public so it can be reused to
parse proto files

**Which issues(s) does this PR fix?**

Fixes #

**Other notes for review**