prometheus: Add directives needed for prometheus and alertmanager (#2032)
This is enough to build promtool, prometheus, amtool and alertmanager
using bazel. The code includes generated gogoproto code and requires to
disable proto handling.

<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

> Other

**What package or component does this PR mostly affect?**

internal/bzlmod

**What does this PR do? Why is it needed?**

Prometheus and Alertmanager have generated gogoproto code that needs to
be built. This makes it easy to use amtool/promtool from non-root
modules.

**Which issues(s) does this PR fix?**


**Other notes for review**