Stop using Label.workspace_name (#2033)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

Other

**What package or component does this PR mostly affect?**

internal/bzlmod

**What does this PR do? Why is it needed?**

According to Bazel's documentation it is deprecated:

https://bazel.build/rules/lib/builtins/Label#workspace_name

We should use Label.repo_name instead.

**Which issues(s) does this PR fix?**

I am working on an analysis tool that is capable of parsing BUILD/*.bzl
files. It currently fails to process these files due to it not
supporting Label.workspace_name. Instead of adding support for it to my
tool, I think I'll just use the occasion to fix upstream rules.

**Other notes for review**