fix: attr.val is not updated when setting multiple times (#2046)
**What type of PR is this?**

Bug fix

**What package or component does this PR mostly affect?**

rule

**Which issues(s) does this PR fix?**

The attr.val field was introduced in the commit a2a604d as part of the
R`ule.SetAttr` implementation. However, there is a bug where updating
`attr.val` does not correctly propagate the change to
`r.attrs[key].val`. The issue stems from Go's map value semantics. When
a value is retrieved from a map, it is returned as a copy, not a
reference. In the current implementation:

```go
if attr, ok := r.attrs[key]; ok {
    attr.expr.RHS = rhs // Fine because RHS is a reference
    attr.val = value  // Modifies the copy, not the original value in the map
}
```
As a result, changes to attr.val are not reflected in r.attrs[key].val.


**Related issue**

https://github.com/bazel-contrib/bazel-gazelle/issues/2045