walk, language: support lazy indexing (#2074)
**What type of PR is this?**

> Feature

**What package or component does this PR mostly affect?**

> walk
> language

**What does this PR do? Why is it needed?**

A language extension may now return a list of additional directories
to index through GenerateResult.RelsToIndex. Gazelle visits these
directories and adds libraries to the dependency resolution index.
The build files in these directories are not updated unless they
would be otherwise.

The -index flag is still enabled by default, but in addition to
"all" (true), "none" (false), it can now be set to "lazy". In lazy mode,
Gazelle only indexes directories listed on the command line, their
subdirectories (if the -r flag is enabled), and directories returned
by extensions through RelsToIndex. This avoids the need to index
the entire repo on each run of Gazelle.

**Which issues(s) does this PR fix?**

For #1891

**Other notes for review**

Part 3/4. #2073 is the previous PR.