language/go: support lazy indexing with go_search directive (#2075)
**What type of PR is this?**

> Feature

**What package or component does this PR mostly affect?**

> language/go

**What does this PR do? Why is it needed?**

The Go extension now supports the "# gazelle:go_search" directive.
This accepts two arguments: a relative directory path, and an
optional import path prefix for packages in that directory.

When lazy indexing is enabled with -index=lazy, the Go extension
returns a directory for each import through
language.GenerateResult.RelsToIndex. This causes Gazelle to index
libraries in directories it might not otherwise visit.

**Which issues(s) does this PR fix?**

Fixes #1891

**Other notes for review**

Part 4/4. #2074 is the previous PR.

---------

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>