Trim BCR versions before comparing them to Go versions (#2066)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

Bug fix

**What package or component does this PR mostly affect?**

`go_deps`

**What does this PR do? Why is it needed?**

Otherwise a BCR version of 1.2.3.bcr.1 will always show a warning when
the version in `go.mod` is 1.2.3.

**Which issues(s) does this PR fix?**


**Other notes for review**