go.mod 'tool' directive may refer to a package at module-root (#2115)
**What type of PR is this?**
Bug fix

**What package or component does this PR mostly affect?**
language/go

**What does this PR do? Why is it needed?**
Some tools have a `main.go` that is at the root of the repo, instead of
nested inside some package. Adding them as a `tool` directive does not
result in them being treated as a direct dependency, so the `use_repo`
will not expose them. Example:
https://github.com/oasdiff/oasdiff/blob/main/main.go

**Which issues(s) does this PR fix?**

Fixes #

**Other notes for review**