walk: expose cached GetDirInfo (#2141)
**What type of PR is this?**

> Feature

**What package or component does this PR mostly affect?**

> walk

**What does this PR do? Why is it needed?**

Adds `walk.GetDirInfo`, a function that exposes the internal I/O cache
used by
`Walk` and `Walk2`. This function does no I/O on its own. It may only be
called
during `Walk` or `Walk2` on directories that have been loaded.

This is meant to help `resolve.Resolver.Imports` implementations resolve
globs.
In C++ for example, the returned imports are based on the `hdrs`
attribute, which
for hand-written rules is likely to be a `glob`. Providing access to
`RegularFiles` and
`GenFiles` lets us expand the `glob` without additional I/O while
respecting `exclude`
and other directives.

**Which issues(s) does this PR fix?**

Fixes #2132

**Other notes for review**