rule: add MatchGlobExpr for parsing glob expressions (#2142)
**What type of PR is this?**

> Feature

**What package or component does this PR mostly affect?**

> rule

**What does this PR do? Why is it needed?**

Adds `MatchGlobExpr`, a function for parsing a `GlobExpr` out of the
syntax tree.
This will be used by the C++ extension's `resolve.Resolver.Imports`
implementation.
It's common for hand-written `cc_library` targets to set the `hdrs`
attribute to a `glob`
expression. We want to be able to expand that `glob` to be able to list
the import
strings for the library.

**Which issues(s) does this PR fix?**

Fixes #2133

**Other notes for review**

As a drive-by fix, `GlobValue.BzlExpr` now uses `Ident` instead of
`LiteralExpr`, which
is for numeric constants.