fix FixLoads issue with list of symbols (#2147)
<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, please read CONTRIBUTING.md and sign the
CLA
   first. We cannot review code without a signed CLA.
2. Please file an issue *first*. All features and most bug fixes should
have
an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
be updated. If you're unsure what to update, send the PR, and we'll
discuss
   in review.
-->

**What type of PR is this?**

Bug fix

**What package or component does this PR mostly affect?**
merger

**What does this PR do? Why is it needed?**
FixLoads does not work for a list expression on the RHS of an assigment
containing symbols.

**Which issues(s) does this PR fix?**
This addresses #2146, but doesn't really solve the underlying issue.

**Other notes for review**