walk: allow GetDirInfo to be called in directories not visited yet (#2153)
**What type of PR is this?**

> Bug fix

**What package or component does this PR mostly affect?**

> walk

**What does this PR do? Why is it needed?**

GetDirInfo was intended to help expand glob expressions in
resolve.Resolver.Imports implementations. globs may cover
subdirectories,
especially when gazelle:generate_mode is used, so GetDirInfo must
be able to read contents of subdirectories that may not have been
visited yet.

Previously, the contract of GetDirInfo did not allow visiting
directories that hadn't been visited yet, but it would still sometimes
succeed walker.populateCache won the race. This change fully
allows GetDirInfo to be called in directories that haven't been
visited yet.

**Which issues(s) does this PR fix?**

Fixes #2132

**Other notes for review**

This PR is based on #2152. I tried to split them for easier review, but
GitHub doesn't let me set the base to a branch in a forked repo, so I'm
sorry if this makes things more difficult.