Extract either Module/Workspace RepoName in `CommonConfigurer` (#2157)
**What type of PR is this?**

Bug fix

**What package or component does this PR mostly affect?**

cmd/gazelle 
config

**What does this PR do? Why is it needed?**

`Config.RepoName` was never setup when using `MODULE.bazel`. 
With this change we now always try to extract repositoryName from either
`module(name = ...)` in MODULE.bazel or `workspace(name = ...) in case
of workspace files.

The logic previously defined in `cmd/gazelle` was moved to
`CommonConfigurer` as this is the place that should be responsible for
setup of RepoName.

**Which issues(s) does this PR fix?**

No issue in this repo. I want to use have special logic in based on
`Config.RepoName` in `gazelle_cc`

**Other notes for review**

The hack (for mapping of google_apis) was preserved in the new
location`. I don't know if it can be removed, or if the mapping should
happen in the `cmd/gazelle` instead.

Signed-off-by: Wojciech Mazur <wmazur@virtuslab.com>